updating code

+++ blog_view.php
@@ -41,10 +41,12 @@
         <div class="blog-header">
             <H1><?php echo $blog['blog_title'] ?></H1>
         </div>
-        <div class="container edit-btn">
-            <a class="btn btn-success" href="update.php?update_id=<?php echo $blog['blog_id'] ?>">Edit</a>
-            <a class="btn btn-danger" href="action.php?delete_id=<?php echo $blog['blog_id'] ?>">Delete</a>
-        </div>
+        <?php if (isset($_SESSION['user_id'])){ ?>
+            <div class="container edit-btn">
+                <a class="btn btn-success" href="update.php?update_id=<?php echo $blog['blog_id'] ?>">Edit</a>
+                <a class="btn btn-danger" href="action.php?delete_id=<?php echo $blog['blog_id'] ?>">Delete</a>
+            </div>
+        <?php } ?>
         <div class="blog-content mt-5 mb-5 container">
             <h3>Introduction</h3>
             <p><?php echo $blog['introduction'] ?></p>
